requirements.txt
google-genai
pydantic
python-dotenv
future
reportlab
